+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Feature: Update `Title` component and add new doc.
 - Feature: Add `RocketIcon` and `RocketCircleIcon` components.
 - Feature: Create `ProgressRing` component.
+- Fix: Modernize `Video` component.
 - Fix: Use `styled-components` instead of `Radium` for `EmbedPlayer` component.
 - Doc: Add a doc for style `utilities`.
 - Feature: Add pagination position `bottom-left` and `bottom-right` on Carousel.